split search types
